Memory LLM/pyproject.toml
[build-system]
requires = ["setuptools>=70.0,<77.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mem-llm"
version = "2.3.2"
description = "Privacy-first, memory-enabled AI assistant with workflow engine, knowledge graph, multi-agent systems, multi-backend LLM support (Ollama, LM Studio), vector search, and analytics - 100% local and production-ready"
readme = {file = "README.md", content-type = "text/markdown"}
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "C. Emre Karataş", email = "karatasqemre@gmail.com"}
]
keywords = ["llm", "ai", "memory", "agent", "multi-agent", "chatbot", "ollama", "lmstudio", "multi-backend", "local", "privacy", "vector-search", "chromadb", "response-metrics", "semantic-search", "quality-analytics", "embedding", "streaming", "function-calling", "tools", "async", "validation", "workspace", "collaboration"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "requests>=2.31.0",
    "pyyaml>=6.0.1",
    "click>=8.1.0",
    "aiohttp>=3.9.0",  # Async tools support (v2.1.0+)
    "networkx>=3.0",  # Knowledge Graph support (v2.3.0+)
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.7.0",
    "flake8>=6.1.0",
]
web = [
    "flask>=3.0.0",
    "flask-cors>=4.0.0",
]
api = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",  # Includes WebSocket support
    "websockets>=12.0",  # WebSocket library
]
postgresql = [
    "psycopg2-binary>=2.9.9",
]
mongodb = [
    "pymongo>=4.6.0",
]
databases = [
    "psycopg2-binary>=2.9.9",
    "pymongo>=4.6.0",
]
all = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.7.0",
    "flake8>=6.1.0",
    "flask>=3.0.0",
    "flask-cors>=4.0.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "websockets>=12.0",
    "psycopg2-binary>=2.9.9",
    "pymongo>=4.6.0",
]

[project.urls]
Homepage = "https://github.com/emredeveloper/Mem-LLM"
"Bug Reports" = "https://github.com/emredeveloper/Mem-LLM/issues"
Source = "https://github.com/emredeveloper/Mem-LLM"

[project.scripts]
mem-llm = "mem_llm.cli:main"
mem-llm-web = "mem_llm.web_launcher:main"

[tool.setuptools]
packages = [
    "mem_llm",
    "mem_llm.clients",
    "mem_llm.web_ui",
    "mem_llm.workflow",
    "mem_llm.memory",
    "mem_llm.memory.hierarchy",
    "mem_llm.memory.graph",
    "mem_llm.multi_agent",
    "mem_llm.analytics"
]

[tool.setuptools.package-data]
mem_llm = [
    "config.yaml.example",
    "web_ui/*.html",
    "web_ui/*.css",
    "web_ui/*.js",
    "web_ui/README.md"
]

# ============================================================================
# Code Quality & Testing Tools Configuration
# ============================================================================

[tool.pytest.ini_options]
minversion = "7.4"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-ra",                    # Show summary of all test outcomes
    "--strict-markers",       # Treat unregistered markers as errors
    "--strict-config",        # Treat config errors as errors
    "--showlocals",          # Show local variables in tracebacks
    "-v",                    # Verbose output
    "--tb=short",            # Shorter traceback format
    "--cov=mem_llm",         # Coverage for mem_llm package
    "--cov-report=term-missing",  # Show missing lines in terminal
    "--cov-report=html",     # Generate HTML coverage report
    "--cov-report=xml",      # Generate XML coverage report (for CI)
    "--cov-branch",          # Measure branch coverage
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "api: API endpoint tests",
    "e2e: End-to-end tests",
    "slow: Tests that take a long time",
    "asyncio: Async tests",
]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["mem_llm"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
    "*/site-packages/*",
    "*/web_ui/*",  # Exclude web UI static files
]
branch = true

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[tool.coverage.html]
directory = "htmlcov"

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | env
  | venv
  | htmlcov
  | \.eggs
  | \.pytest_cache
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip_gitignore = true
skip = [
    ".git",
    ".tox",
    ".venv",
    "venv",
    "env",
    "build",
    "dist",
    "*.egg-info",
]
known_first_party = ["mem_llm"]
known_third_party = [
    "requests",
    "yaml",
    "click",
    "aiohttp",
    "fastapi",
    "uvicorn",
    "pytest",
]

[tool.flake8]
max-line-length = 100
extend-ignore = [
    "E203",  # Whitespace before ':' (conflicts with Black)
    "E266",  # Too many leading '#' for block comment
    "E501",  # Line too long (handled by Black)
    "W503",  # Line break before binary operator
]
exclude = [
    ".git",
    "__pycache__",
    "build",
    "dist",
    ".eggs",
    "*.egg-info",
    ".tox",
    ".venv",
    "venv",
    "env",
]
per-file-ignores = [
    "__init__.py:F401",  # Allow unused imports in __init__.py
]
max-complexity = 10

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false  # Gradually introduce type hints
disallow_incomplete_defs = false
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
show_error_codes = true
show_column_numbers = true
pretty = true

[[tool.mypy.overrides]]
module = [
    "chromadb.*",
    "sentence_transformers.*",
    "psycopg2.*",
    "pymongo.*",
]
ignore_missing_imports = true
